+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2026-02-26
+
+- feat: add external documentation sources support via bartleby.sources manifest config
+
 ## 2026-02-25
 
 - feat: add slides subcommand for RevealJS slideshow rendering